Separate ModuleResolverCache into lib/module-resolver-cache.js
